細かいバグ修正.

+++ Utility/Time.h
@@ -85,7 +85,7 @@ namespace SDX
 			buf += " = ";
 
 			auto diff = std::chrono::system_clock::now() - Single().watch;
-			double count = std::chrono::duration_cast<std::chrono::milliseconds>(diff).count();
+			double count = (double)std::chrono::duration_cast<std::chrono::milliseconds>(diff).count();
 			Drawing::String(座標, Color(255, 255, 255), { buf, count });
 			Single().watch = std::chrono::system_clock::now();
 		}
